requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
distro==1.9.0
dotenv==0.9.9
filelock==3.19.1
Flask==3.1.1
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
ollama==0.5.3
openai==1.100.1
openai-whisper==20250625
pydantic==2.11.7
pydantic_core==2.33.2
pypiwin32==223
python-dotenv==1.1.1
pywin32==311
regex==2025.7.34
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
